requirements.txt
alembic==1.7.7
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0
defusedxml==0.7.1
Flask==2.3.2
Flask-HTTPAuth==4.5.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
httpie==2.6
idna==3.3
importlib-metadata==4.8
itsdangerous==2.0
Jinja2==3.0
Mako==1.1
MarkupSafe==2.0
multidict==5.2
netifaces==0.11.0
psycopg2-binary==2.9.3
Pygments==2.11.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
SQLAlchemy==1.4.32
urllib3==1.26.9
Werkzeug==2.0.3
WTForms==3.0
zipp==3.6
